Convert RSS feed to Atom
When adding authors to rss items, there doesn't seem to be a good standard. Atom does have a good standard however.

This will cause the marketing copy to change, and obviously the feed response, but overall it isn't that crazy and I was able to clean up a bunch of code.