Create unique uri for entry IDs
According to the [atom standard](http://tools.ietf.org/html/rfc4287#page-20) entries in an atom feed should never, ever change. They should also not be the same as the URI (or really even be a URI) to limit confusion as to what they are used for. They are a unique string used for identification.

`pullfeed.co` is hardcoded into for tag uri formula so that even if the app's domain changes, the unique, never changing ids will stay the same. Hopefully.

Here is a blog post for reference: https://web.archive.org/web/20081225232147/http://diveintomark.org/archives/2004/05/28/howto-atom-id