RepositorySerializer can now handle absense of PRs
Before, it was assumed that PR data was being passed into the RepositorySerializer. But if there are no PRs, an empty array is all that will be passed. It now intelligently handles raw repo data.

This will be helpful in the next few commits when I fetch repo data instead.