MixpanelAnalytics for tracking feed requests
The ability to track what feeds are being accessed most often is very useful. The MixpanelAnalytics module can be inclued in any controller and that controller can send an `owner` and `repo` to `requested_feed` which will send it off to Mixpanel. The module is made to be expandable to other types of events that the app may want to track in the future.

This also added an additional request to the FeedsController (the one to mixpanel) that needed to be stubbed out.