Add github access token to requests
The app was hitting the rate limit a bunch, which is great because people are using it, but rate limit errors are not great. To solve this, the app can use an access token and get many more requests per hour (5000 instead of 60).

This branch adds those tokens to the requests.

https://developer.github.com/v3/oauth_authorizations/#create-a-new-authorization