Create SubscriptionParser
This object takes an ActionDispatch::Request and parses out the
subscriber and reader count. For example, if the request has the
User-Agent of "Bloglines/3.0-rho (http://www.bloglines.com; 3
subscribers)", it will return

```ruby { count: 3, subscriber: 'Bloglines' } ```

If there is no user_agent string, the class will make sure it's set to a
blank string. This will make further operations not crash, like regex or
subscriber matching.

When there is no match at all for the user_agent string, it will return
1 instead of crashing.