Convert HTTParty response in the fetching objects
The last commit attempted to grab the parsed response from the HTTParty response object inside of RepositorySerializer, which needed to be done to have the data I expected for the rest of the pull request conversion process. However, it's equally valid to convert it inside of the fetcher objects (RepositoryFetcher, PullRequestFetcher) and doing it this way let me keep the tests the same.

Thanks to @rosmith76 and @asackofwheat