Update Puma config
This is a very similar config to the one I use on my portfolio at edwardloveall.com. The only difference is I don't bind to a socket if I'm in the Development environment. I'm not sure why that was in my portfolio app, and changing it over there might fix some things!